Fix test failure on Windows caused by nil `sys` in mock `FileInfo`
Fix test failure caused by `panic` when checking if a file is hidden in
Windows. The hidden check uses `FileInfo.Sys()`. In tests the object is
a mock and that call returns `nil`. Regardless of mocking, we should
check for `nil` since according to docs that function can return `nil`.

Run `go mod tidy`.

Relates to:
- https://github.com/ipfs/go-ipfs-files/pull/34